Module attiny85_hal::prelude

Re-exports

pub use crate::port::PortExt as _;

Traits

_
Single digital push-pull output pin

_
Single digital input pin

_
Push-pull output pin that can read its output state

_
Output pin that can be toggled

_
Extensions to Result<T, Void>

_
Extensions to Result<Void, E>

_
A collection of methods that are required / used to format a message into a stream.

_embedded_hal_Capture
Input capture

_embedded_hal_Pwm
Pulse Width Modulation

_embedded_hal_PwmPin
A single PWM channel / pin

_embedded_hal_Qei
Quadrature encoder interface

_embedded_hal_adc_OneShot
ADCs that sample on single channels per request, and do so at the time of the request.

_embedded_hal_blocking_delay_DelayMs
Millisecond delay

_embedded_hal_blocking_delay_DelayUs
Microsecond delay

_embedded_hal_blocking_i2c_Read
Blocking read

_embedded_hal_blocking_i2c_Write
Blocking write

_embedded_hal_blocking_i2c_WriteRead
Blocking write + read

_embedded_hal_blocking_rng_Read
Blocking read

_embedded_hal_blocking_serial_Write
Write half of a serial interface (blocking variant)

_embedded_hal_blocking_spi_Transfer
Blocking transfer

_embedded_hal_blocking_spi_Write
Blocking write

_embedded_hal_digital_InputPin	Deprecated
Single digital input pin

_embedded_hal_digital_OutputPin	Deprecated
Single digital push-pull output pin

_embedded_hal_digital_ToggleableOutputPin	Deprecated
Output pin that can be toggled

_embedded_hal_serial_Read
Read half of a serial interface

_embedded_hal_serial_Write
Write half of a serial interface

_embedded_hal_spi_FullDuplex
Full duplex (master mode)

_embedded_hal_timer_CountDown
A count down timer

_embedded_hal_watchdog_Watchdog
Feeds an existing watchdog to ensure the processor isn't reset. Sometimes commonly referred to as "kicking" or "refreshing".

_embedded_hal_watchdog_WatchdogDisable
Disables a running watchdog timer so the processor won't be reset.

_embedded_hal_watchdog_WatchdogEnable
Enables A watchdog timer to reset the processor if software is frozen or stalled.